# src/main/resources/openapi/spec/employees.yaml
openapi: 3.0.0
info:
  description: Provides the specification for the Employees Microservice API
  version: "1.0.0"
  title: EMPLOYEES API

tags:
  - name: Employees
    description: Controller to handle employees

components:
  responses:
    BadRequest:
      description: "Indicates the message body contained one or more errors."
    NotFound:
      description: "Indicates the identified resource does not exist."
    Conflict:
      description: "Indicates that the resource already exist and cannot be created"

  schemas:

    EmployeeRequestDTO:
      type: object
      description: An employee request
      required:
        - name
        - description
      properties:
        name:
          type: string
          description: The employee name
        description:
          type: string
          description: The employee description
        departmentId:
          type: integer
          description: The department Id
        age:
          type: integer
          description: The employee age


    EmployeeResponseDTO:
      type: object
      allOf:
        - $ref: '#/components/schemas/EmployeeRequestDTO'
      description: An response employee object
      properties:
        id:
          type: integer
          description: The employee Id

paths:
  "/employees":
    get:
      tags:
        - Employees
      operationId: getEmployees
      description: Gets all the employees
      parameters:
        - name: page
          in: query
          required: true
          schema:
            type: integer
        - name: size
          in: query
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: The employees were retrieved successfully
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/EmployeeResponseDTO'
        "404":
          $ref: "#/components/responses/NotFound"
    post:
      tags:
        - Employees
      operationId: addEmployee
      description: Add an employee
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EmployeeRequestDTO"
      responses:
        "201":
          description: Employee has been added.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponseDTO'
        "404":
          $ref: "#/components/responses/NotFound"

  "/employees/{employeeId}":
    get:
      tags:
        - Employees
      operationId: getEmployee
      description: Retrieve a employee based on employee Id
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Employee has been retrieved successfully.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponseDTO'
        "404":
          $ref: "#/components/responses/NotFound"
    delete:
      tags:
        - Employees
      operationId: deleteEmployee
      description: Deletes an employee based on the employee id
      parameters:
        - name: employeeId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Employee has been deleted.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/EmployeeResponseDTO'
        "404":
          $ref: "#/components/responses/NotFound"